% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{census_vars}
\alias{census_vars}
\title{Census variables}
\format{
A data frame with 9 rows and 2 variables:
\describe{
\item{\code{variable}}{character Census variable name}
\item{\code{varname}}{character Short name used in zipcode table}
}
}
\usage{
census_vars
}
\description{
Lookup table for ACS variable names
}
\details{
None.
}
\keyword{datasets}
