% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nyc_zips}
\alias{nyc_zips}
\title{NYC Zip Codes by borough}
\format{
A data frame with 211 rows and 5 variables:
\itemize{
\item \code{zip} Zip code (char)
\item \code{borough} Borough name
\item \code{city} City name (USPS)
\item \code{county} County name
\item \code{long_county} County name, New York
\item \code{short_county} One word county  name
}
}
\usage{
nyc_zips
}
\description{
Zip codes and borough names
}
\details{
Sources: https://www.nyc.gov/assets/planning/download/office/data-maps/nyc-population/census2000/sf1p11.xls
and https://worldpopulationreview.com/zips/new-york
}
\keyword{datasets}
