% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nycomplaints}
\alias{nycomplaints}
\title{Complaints to NYC Council Constituent Services}
\format{
A data frame with 341,299 rows and 11 variables:
\describe{
\item{\code{unique_key}}{character Complaint identifier.}
\item{\code{account}}{character NYC Council District that created the casework, this may or not be the district that the constituent lives in.}
\item{\code{opendate}}{date Date the casework is created by staff.}
\item{\code{closedate}}{date Close Date is the date the casework was closed. A casework could be closed because the issue is resolved or because no further action could be taken.}
\item{\code{complaint_type}}{character Broader category designated by staff for the constituent casework. complaint_type corresponds to NYC Council Legislative Committees.}
\item{\code{descriptor}}{character Specific category designated by staff for the constituent casework.}
\item{\code{zip}}{integer Zip code of the constituent.}
\item{\code{borough}}{character Borough of the constituent.}
\item{\code{city}}{character City of the constituent.}
\item{\code{council_dist}}{Council District of the constituent.}
\item{\code{community_board}}{Community Board of the constituent.}

}
}
\usage{
nycomplaints
}
\description{
Complaints data from New York City's CouncilStat
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nycomplaints \cr
   Number of rows \tab 341299 \cr
   Number of columns \tab 11 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 2 \cr
   character \tab 9 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   opendate \tab 0 \tab 1.00 \tab 2015-01-01 \tab 2025-01-09 \tab 2018-06-19 \tab 3296 \cr
   closedate \tab 14142 \tab 0.96 \tab 2015-01-02 \tab 2025-01-09 \tab 2018-11-20 \tab 2686 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   unique_key \tab 0 \tab 1.00 \tab 12 \tab 12 \tab 0 \tab 341194 \tab 0 \cr
   account \tab 0 \tab 1.00 \tab 6 \tab 6 \tab 0 \tab 51 \tab 0 \cr
   complaint_type \tab 7431 \tab 0.98 \tab 3 \tab 25 \tab 0 \tab 59 \tab 0 \cr
   descriptor \tab 20507 \tab 0.94 \tab 1 \tab 56 \tab 0 \tab 1259 \tab 0 \cr
   zip \tab 50624 \tab 0.85 \tab 1 \tab 5 \tab 0 \tab 1370 \tab 0 \cr
   borough \tab 63097 \tab 0.82 \tab 5 \tab 13 \tab 0 \tab 5 \tab 0 \cr
   city \tab 35555 \tab 0.90 \tab 1 \tab 49 \tab 0 \tab 1465 \tab 0 \cr
   council_dist \tab 60210 \tab 0.82 \tab 4 \tab 12 \tab 0 \tab 156 \tab 0 \cr
   community_board \tab 72772 \tab 0.79 \tab 1 \tab 16 \tab 0 \tab 99 \tab 0 \cr
}


Produced from NYC Open Data Initiative data. The dataset comes from CouncilStat, which is used by many NYC Council district offices to enter and track constituent cases that can range from issues around affordable housing, to potholes and pedestrian safety. This dataset aggregates the information that individual staff have input. However, district staffs handle a wide range of complex issues. Each offices uses the program differently, and thus records cases, differently and so comparisons between accounts may be difficult. Not all offices use the program. For more info - \url{http://labs.council.nyc/districts/data/}
}
\references{
\url{https://data.cityofnewyork.us/City-Government/NYC-Council-Constituent-Services/9hzi-kbqb},\url{http://labs.council.nyc/districts/data/}
}
\author{
Kieran Healy
}
\keyword{datasets}
