% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nyzip_demog}
\alias{nyzip_demog}
\title{Demographic data for NYC zip codes}
\format{
A data frame with 68 rows and 10 variables:
\describe{
\item{\code{zip}}{character Zip code}
\item{\code{population}}{double Population count: B01001_001}
\item{\code{white_alone}}{double B02001_002}
\item{\code{black_alone}}{double B02001_003}
\item{\code{asian_alone}}{double B02001_005}
\item{\code{two_or_more_races}}{double B02001_008}
\item{\code{hispanic}}{double B03003_001}
\item{\code{nonhispanic_white}}{double B03002_003}
\item{\code{nonhispanic_black}}{double B03002_004}
\item{\code{med_hhinc}}{double B19013_001}
}
}
\usage{
nyzip_demog
}
\description{
USPS zip codes with corresponding census ZCTAs
}
\details{
USPS zip code list with corresponding ZCTAs; demographic data from the ACS
}
\keyword{datasets}
