% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nycomplaints}
\alias{nycomplaints}
\title{Complaints to NYC Council Constituent Services}
\format{
A data frame with 289,765 rows and 11 variables:
\describe{
\item{\code{unique_key}}{character Complaint identifier.}
\item{\code{account}}{character NYC Council District that created the casework, this may or not be the district that the constituent lives in.}
\item{\code{opendate}}{date Date the casework is created by staff.}
\item{\code{closedate}}{date Close Date is the date the casework was closed. A casework could be closed because the issue is resolved or because no further action could be taken.}
\item{\code{complaint_type}}{character Broader category designated by staff for the constituent casework. complaint_type corresponds to NYC Council Legislative Committees.}
\item{\code{descriptor}}{character Specific category designated by staff for the constituent casework.}
\item{\code{zip}}{integer Zip code of the constituent.}
\item{\code{borough}}{character Borough of the constituent.}
\item{\code{city}}{character City of the constituent.}
\item{\code{council_dist}}{Council District of the constituent.}
\item{\code{community_board}}{Community Board of the constituent.}

}
}
\usage{
nycomplaints
}
\description{
Complaints data from New York City's CouncilStat
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab nycomplaints \cr
   Number of rows \tab 289765 \cr
   Number of columns \tab 11 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 2 \cr
   character \tab 9 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   opendate \tab 0 \tab 1.00 \tab 2015-01-01 \tab 2025-01-09 \tab 2018-06-13 \tab 3287 \cr
   closedate \tab 12046 \tab 0.96 \tab 2015-01-02 \tab 2025-01-09 \tab 2018-11-16 \tab 2665 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   unique_key \tab 0 \tab 1.00 \tab 12 \tab 12 \tab 0 \tab 289693 \tab 0 \cr
   account \tab 0 \tab 1.00 \tab 6 \tab 6 \tab 0 \tab 51 \tab 0 \cr
   complaint_type \tab 5969 \tab 0.98 \tab 3 \tab 25 \tab 0 \tab 58 \tab 0 \cr
   descriptor \tab 17970 \tab 0.94 \tab 1 \tab 56 \tab 0 \tab 1182 \tab 0 \cr
   zip \tab 0 \tab 1.00 \tab 5 \tab 5 \tab 0 \tab 1196 \tab 0 \cr
   borough \tab 21893 \tab 0.92 \tab 5 \tab 13 \tab 0 \tab 5 \tab 0 \cr
   city \tab 1143 \tab 1.00 \tab 1 \tab 36 \tab 0 \tab 1144 \tab 0 \cr
   council_dist \tab 15284 \tab 0.95 \tab 4 \tab 12 \tab 0 \tab 155 \tab 0 \cr
   community_board \tab 21814 \tab 0.92 \tab 1 \tab 16 \tab 0 \tab 98 \tab 0 \cr
}


Produced from NYC Open Data Initiative data. The dataset comes from CouncilStat, which is used by many NYC Council district offices to enter and track constituent cases that can range from issues around affordable housing, to potholes and pedestrian safety. This dataset aggregates the information that individual staff have input. However, district staffs handle a wide range of complex issues. Each offices uses the program differently, and thus records cases, differently and so comparisons between accounts may be difficult. Not all offices use the program. For more info - \url{http://labs.council.nyc/districts/data/}
}
\references{
\url{https://data.cityofnewyork.us/City-Government/NYC-Council-Constituent-Services/9hzi-kbqb},\url{http://labs.council.nyc/districts/data/}
}
\author{
Kieran Healy
}
\keyword{datasets}
